fix(scripts/train): 修复训练日志图片数据类型不匹配问题
将拼接后的图像数组转换为uint8类型，确保图像日志可以正确显示和保存

+++ scripts/train.py
@@ -370,7 +370,7 @@ def main(config: _config.TrainConfig):
 
     # Log images from first batch to sanity check.
     images_to_log = [
-        trackio.Image(np.concatenate([np.array(img[i]) for img in batch[0].images.values()], axis=1))
+        trackio.Image(np.concatenate([np.array(img[i]) for img in batch[0].images.values()], axis=1).astype(np.uint8))
         for i in range(min(5, len(next(iter(batch[0].images.values())))))
     ]
     trackio.log({"camera_views": images_to_log}, step=0)
